Add fix for folders being added to a watched folder
In FolderReader, if a folder is added to a watched folder it is incorrectly
marked as a file.

Tests and a fix have been added to correct the behaviour.
If a folder is added then the folder is registered and watched, if recursive is set to true,
and otherwise it is ignored.