Baleen Knowledge Representation
This commit includes improvements to existing knowledge represenation, new
features for graph represenation and an alternative approach the current
Mongo and Elasticsearch consumers tuned for analysis.

To support graphs, we first create a full represenation of the document
annotations as a graph, called the DocumentGraph. Then particular
represenations can be derived by transforming that graph. The EntityGraph
is such a transformation that concentrates on the (coreferenced) entities
(or ReferenceTarget in Baleen).

* LocationElasticsearch indexes Location annotaitons per document
* TemporalElasticsearch indexes Temporal annotations per document
* MongoRealations gives a separate representation of Relation annotations
* MongoEvents gives a separate representation of Even annotations
* analysis.Mongo is an alternative represenation in Mongo for analysis applications
* analysis.Elasticsearch is an alternative represenation in Elasticsearch for analysis applications
* baleen-graph allows the represeantion of the data as a (Tinkerpop) graph that can
be output to file or supported graph databases.
* baleen-graph-neo4j adds specific support for neo4j using the bolt protocol
* baleen-rdf allows the output of the graphs in RDF formats and tripple
stores using