Baleen Relations
This commit adds additional relation extraction features.

We add distance measures to Relation and regenerate the type
(this changes more than expected, but have accepted as generated)

SentenceRelationshipAnnotator and DocumentRelationshipAnnotator
are simple annotators the can be used to baseline the relation content and
and investigate relation structures in a corpus but not intended for general use.

A set of pattern-based annotator allow regular expression based relations
mixing entities or parts of speech into the patterns.

And significant support for extraction using dependency patterns is added.
This is then used to add the ReNoun system for noun based relation extraction.
This requires setup for the corpus.