Baleen Triage
We collect some current annotators associated with document triage
into the `triage` namespace and contribute annotators to assign a
document date form the title, create document summaries and compute the
Shannon Entropy of the document to inform prioritisation.

We also integrate the Mallet library for document classification.
This supports a number of different algorithms for learning from labelled
documents, learning by suggestion and a Latent Dirichlet approach for when
no labels data exists.