Baleen Scalability (#69)
* Baleen Scalability

This commit allows the scaling of Baleen pipelines.
It works by having a reading pipeline that uses a consumer to write the
whole jCas to a transport system and other pipelines read from that
transport system. This allows scaling within the same Baleen and across
multiple instances. There is currently no coordination, this is still
done by configuring the pipelines appropriately for your scaling
requirements. To simplify this a multiplicity property can be set on the
pipline to create multiple instances. This option is available in the
Baleen configuration and the API.

The following transport systems are supported:

* In Memory - for testing and in server transports
* ActiveMQ (NB all ActiveMQ dependencies move under baleen-activemq)
* Kafka
* RabbitMQ
* Redis

Testing

There are mock resources used to test the transports.
We run some tests using embedded services and for all, we provide a docker
based integration test using 'test-containers' these require docker and can
be run using:

```
mvn integration-test -Dskip.unit.tests=true -Dskip.integration.tests=false
```

* Minor updates to scalability sprint

Reverts incorrectly change copyright
Improves doc on memory transports