Baleen Entity Linking (#73)
* Baleen Entity Linking

In this commit, we provide a framework for entity linking and an initial
implementation. Entity linking is the identification of entities in a document
against an external data source.

We provide example candidate suppliers from DBpedia and a simple Mongo
document based supplier. A general matching algorithm using a bag of words
approach is also supplied but particular situations will benefit from tuned
implementations.

* Rename Propper to Proper

* Simplifying Information Collector and Ranker test