Force the use of liblinear version 1.94
This stops the close stream issue when using OdinParser.
It remains possible that this introduces issue for the maltparser.
In particular, it is lileky to leave a model file reader unclosed,
due to the change in behaviour between the two versions, though
it's not clear that this particular code path is used in Baleen.

We do know that all the test pass and manual testing has not produced an issue.

Fixes #80