Update the entity graph files
This fixes the EntityGraphFileTest test failures on jdk-11 by adding
two versions of the expected serialized files.

It is possible this will fail on jdk 9 or 10 now as these have not been tested.

There would seem to be an implementation difference between Properties
that affects the serialisation order but should not affect teh ability to
deserialize in different java platforms.